% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown_pdf_document.R
\name{AssignmentsRMD}
\alias{AssignmentsRMD}
\title{Customized R Markdown Document}
\usage{
AssignmentsRMD(...)
}
\arguments{
\item{...}{Passed on to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
\code{AssignmentsRMD()} is the main function of the AssignmentsRMD package. Use as
\code{output: AssignmentsRMD::AssignmentsRMD} in the R Markdown YAML header.
}
