% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMDs4NMU.R
\name{tests}
\alias{tests}
\title{tests}
\usage{
tests(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{TRUE}}
}
\value{
R Markdown output to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
A format for LaTeX typeset tests that use the exam
document class via rmarkdown.
}
